feat: :sparkles: Rewrote some components and pages to make them unique
