+++ src/mangaInterfaces/chapterList.ts
