+++ src/mangaInterfaces/chapters.ts
