+++ src/mangaInterfaces/covers.ts
