+++ src/mangaInterfaces/manga.ts
