+++ src/mangaInterfaces/mangainfo.ts
